+++ Changelog.md
@@ -7,6 +7,7 @@ Features:
  * Introduce ``.transfer(value)`` for sending Ether.
  * Code generator: Support ``revert()`` to abort with rolling back, but not consuming all gas.
  * Inline assembly: Support ``revert`` (EIP140) as an opcode.
+ * Parser: Support scientific notation in numbers (e.g. ``2e8`` and ``200e-2``).
  * Type system: Support explicit conversion of external function to address.
  * Type system: Warn if base of exponentiation is literal (result type might be unexpected).